feat : auth 생성 추가
 - generateToken 추가
 - AuthController 추가
