--- buretteAuth/.gradle/buildOutputCleanup/cache.properties
@@ -1,2 +1,0 @@
-#Mon Mar 29 01:42:52 KST 2021
-gradle.version=6.3
