+++ buretteAuth/settings.gradle
@@ -1,2 +1,1 @@
 rootProject.name = 'buretteAuth'
-
