+++ buretteAuth/src/main/java/com/takeon/burette/BuretteAuthApplication.java
@@ -1,0 +1,13 @@
+package com.takeon.burette;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BuretteAuthApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(BuretteAuthApplication.class, args);
+	}
+
+}
