+++ buretteAuth/src/main/java/com/takeon/burette/dto/TokenRequest.java
@@ -1,0 +1,13 @@
+package com.takeon.burette.dto;
+
+public class TokenRequest {
+    private String id;
+
+    public TokenRequest(String id) {
+        this.id = id;
+    }
+
+    public String getId() {
+        return id;
+    }
+}
