+++ buretteAuth/src/test/java/com/takeon/burette/BuretteAuthApplicationTests.java
@@ -1,0 +1,13 @@
+package com.takeon.burette;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BuretteAuthApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
